--- packages/twenty-front/src/pages/settings/data-model/utils/__tests__/compute-metadata-name-from-label.test.ts
@@ -1,9 +1,0 @@
-import { computeMetadataNameFromLabel } from '~/pages/settings/data-model/utils/compute-metadata-name-from-label.utils';
-
-describe('computeMetadataNameFromLabel', () => {
-  it('computes name for label with non-latin char', () => {
-    const label = 'λλλ!';
-
-    expect(computeMetadataNameFromLabel(label)).toEqual('lll');
-  });
-});
