+++ packages/twenty-front/src/pages/settings/data-model/utils/computeMetadataNameFromLabel.ts
@@ -1,0 +1,9 @@
+import { computeMetadataNameFromLabelOrThrow } from '~/pages/settings/data-model/utils/computeMetadataNameFromLabelOrThrow';
+
+export const computeMetadataNameFromLabel = (label: string): string => {
+  try {
+    return computeMetadataNameFromLabelOrThrow(label);
+  } catch {
+    return '';
+  }
+};
