+++ packages/twenty-front/src/pages/settings/data-model/utils/computeOptionValueFromLabel.ts
